init client space

+++ RC/Backend/AuthenticationService/src/main/java/com/AuthenticationService/AuthenticationService/Model/user.java
@@ -15,7 +15,7 @@ public class user {
   @Id
   @GeneratedValue(strategy = GenerationType.SEQUENCE)
   @Column(name = "id", nullable = false)
-  private float id;
+  private Long id;
   private String nom;
   private String prenom;
   private String email;
